Introduce formal verification for loads and stores
This commit adds two new symbolic test for verifying the decoding process of the loads and stores. The process found one bug which was fixed in a recent previous commit.
