+++ model_checking/Cargo.toml
@@ -8,8 +8,6 @@ miralis = { path = "../src", features = ["userspace"] }
 sail_prelude = { path = "sail_prelude"}
 sail_model = { path = "sail_model"}
 sail_decoder = { path = "sail_decoder"}
-sail_decoder_load = { path = "sail_decoder_load"}
-sail_decoder_store  = { path = "sail_decoder_store"}
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(kani)'] }
