--- model_checking/sail_decoder_store/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "sail_decoder_store"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "lib.rs"
-
-[dependencies]
-sail_prelude = { path = "../sail_prelude" }
-sail_model = {path = "../sail_model"}
