--- model_checking/sail_decoder_store/readme.md
@@ -1,4 +1,0 @@
-# Sail decoder
-
-The files in this folder are generated automatically.
-They are extracted from the [RISC-V Sail model](https://github.com/riscv/sail-riscv) and we use them to prove check that Miralis implements the RISC-V specification correctly.
